backend/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.17
redis>=5.0.0
celery[redis]>=5.4.0
alembic>=1.14.0
pillow>=11.0.0
boto3>=1.35.0
pytest>=8.0.0
httpx>=0.27.0
